Init script

See merge request richard.chudy/dashboard!26

+++ core/src/main/java/com/globallogic/dashboard/sprint/SprintData.java
@@ -14,9 +14,9 @@ public class SprintData implements Serializable {
     @Id
     private Long id;
 
-    private float storyPointsTaken;
+    private Double storyPointsTaken;
 
-    private float storyPointsClosed;
+    private Double storyPointsClosed;
 
     @ManyToOne
     private Team team;
@@ -37,19 +37,19 @@ public class SprintData implements Serializable {
         this.id = id;
     }
 
-    public float getStoryPointsTaken() {
+    public Double getStoryPointsTaken() {
         return storyPointsTaken;
     }
 
-    public void setStoryPointsTaken(float storyPointsTaken) {
+    public void setStoryPointsTaken(Double storyPointsTaken) {
         this.storyPointsTaken = storyPointsTaken;
     }
 
-    public float getStoryPointsClosed() {
+    public Double getStoryPointsClosed() {
         return storyPointsClosed;
     }
 
-    public void setStoryPointsClosed(float storyPointsClosed) {
+    public void setStoryPointsClosed(Double storyPointsClosed) {
         this.storyPointsClosed = storyPointsClosed;
     }
